backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==1.4.54
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
aiosqlite==0.19.0